pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Resh"]
readme = "README.md"

[tool.poetry.dependencies]
#python = "^3.9"
python = ">=3.9,<3.11"
jupyter = "^1.0.0"
pandas = "1.5.3"
pyarrow = "^13.0.0"
fastparquet = "^2023.8.0"
matplotlib = "^3.8.0"
plotly = "^5.18.0"
scikit-learn = "^1.3.2"
xgboost = "^2.0.3"
lightgbm = "^4.3.0"
geopy = "^2.4.1"
optuna = "^3.5.0"
joblib = "^1.3.2"
python-dotenv = "^1.0.1"
hopsworks = "^3.4.4"
setuptools = "^69.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
